Continued implementing CENormPriorsSubtaskMax . Implemented "set to bit quad" in SRSimd .

+++ ProbQA/SRPlatform/BucketerSubtaskSum.cpp
@@ -28,7 +28,9 @@ template<> void BucketerSubtaskSum<SRDoubleNumber>::Run() {
     total = SRSimd::HorizAddStraight(total, SumColumn(i)._comps);
   }
   if (isAtPartial) {
-    const __m256d mask = _mm256_castsi256_pd(SRSimd::SetLsb1(task._nValid * SRNumTraits<double>::_cnTotalBits));
+    //TODO: refactor to a table + _mm_cvtsi32_si128()
+    const __m256d mask = _mm256_castsi256_pd(SRSimd::SetLsb1(
+      uint16_t(task._nValid) * SRNumTraits<double>::_cnTotalBits));
     total = SRSimd::HorizAddStraight(total, _mm256_and_pd(mask, SumColumn(task._iPartial)._comps));
   }
   _pBs->_pWorkerSums[_iWorker]._comps = total;
